+++ services/backend/src/utils/featureFlag.js
@@ -3,6 +3,8 @@ const database = require('../database');
 const DEFAULT_VALUES = {
   sms_rate_mm: 1,
   sms_rate_sinch: 1,
+  dummy_max_tracings: 10,
+  dummy_max_traces: 20,
 };
 
 const get = async key => {